refactor: split config tests into smaller chunks
